[CBRD-25924] revised timezone results temporarily due to incorrect DST changes in time_format()
Refer to http://jira.cubrid.org/browse/CBRD-25924

This issue occurred because the date value passed to the time_format function was ignored. Specifically, instead of being processed based on the timezone of "2024-09-01," it was handled according to the system date's timezone.

Due to this issue, these test cases are currently unstable because the processing follows the system date's timezone, causing the "%TZH" result to change during the start and end of Daylight Saving Time (DST). However, due to time constraints, we will temporarily update the expected output accordingly. Once CBRD-25924 is resolved, we will review and update the expected results based on the new output.